added metadata interface to handle non-specific metadata
