+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/fiberphotometryinterface.py
@@ -51,19 +51,10 @@ class FiberPhotometryInterface(BaseDataInterface):
         session_metadata = session_metadata[self.source_data["session_uuid"]]
         subject_metadata = subject_metadata[session_metadata["subject_id"]]
 
-        # Session metadata
-        metadata["NWBFile"]["session_description"] = session_metadata["session_description"]
-        metadata["NWBFile"]["session_start_time"] = session_metadata["session_start_time"]
-        metadata["NWBFile"]["identifier"] = self.source_data["session_uuid"]
-        metadata["NWBFile"]["session_id"] = self.source_data["session_uuid"]
         metadata["FiberPhotometry"]["reference_max"] = session_metadata["reference_max"]
         metadata["FiberPhotometry"]["signal_max"] = session_metadata["signal_max"]
         metadata["FiberPhotometry"]["signal_reference_corr"] = session_metadata["signal_reference_corr"]
         metadata["FiberPhotometry"]["snr"] = session_metadata["snr"]
-
-        # Subject metadata
-        metadata["Subject"]["subject_id"] = session_metadata["subject_id"]
-        metadata["Subject"]["sex"] = subject_metadata["sex"]
         metadata["FiberPhotometry"]["area"] = subject_metadata["photometry_area"]
 
         return metadata
