+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/nwbconverter.py
@@ -4,12 +4,14 @@ from .behaviorinterface import BehaviorInterface
 from .fiberphotometryinterface import FiberPhotometryInterface
 from .optogeneticinterface import OptogeneticInterface
 from .behavioralsyllableinterface import BehavioralSyllableInterface
+from .metadatainterface import MetadataInterface
 
 
 class NWBConverter(NWBConverter):
     """Primary conversion class."""
 
     data_interface_classes = dict(
+        Metadata=MetadataInterface,
         Behavior=BehaviorInterface,
         BehavioralSyllable=BehavioralSyllableInterface,
         FiberPhotometry=FiberPhotometryInterface,
